+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * [Feature] Using native DuckDB `.df()` method when using `autopandas` 
 
 * [Doc] documenting `%sqlcmd tables`/`%sqlcmd columns`
+* [Feature] Better error messages when function used in plotting API unsupported by DB driver (#159)
 
 ## 0.7.4 (2023-04-28)
 No changes